Información detallada de datasets / Detailed dataset information

Muestra información detallada de los datasets disponibles, incluyendo descripción de variables y recursos.

Shows detailed information about available datasets, including variable descriptions and resources.

Usage

vp_dataset_info(dataset = NULL)

Arguments

dataset

Nombre del dataset ("malaria", "dengue", etc.) o NULL para todos. / Dataset name ("malaria", "dengue", etc.) or NULL for all.

Value

Lista con información detallada / List with detailed information

Examples

# Información de malaria / Malaria information
vp_dataset_info("malaria")
#> $dataset
#> [1] "malaria"
#> 
#> $titulo
#> [1] "Vigilancia epidemiológica de Malaria"
#> 
#> $descripcion
#> [1] "<h1>Dataset de notificación de casos de malaria&nbsp;al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <strong>malaria</strong><b>&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483 son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>Se cuentan con dos dataset el primero contempla una serie histórica del 2000 al 2008, con datos desagregados por ano,semana,departamento,provincia,distrito,ubigeo,falciparum,vivax.</p>\n\n<p>El dataset incluye las siguientes variables:</p>\n\n<table border=\"0\" cellpadding=\"0\" cellspacing=\"0\" width=\"671\">\n\t<tbody>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\" width=\"90\">CAMPO</td>\n\t\t\t<td width=\"80\">TIPO</td>\n\t\t\t<td width=\"231\">DENOMINACIÓN</td>\n\t\t\t<td width=\"270\">VALORES</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td colspan=\"2\">Código del lugar probable de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">falciparum</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Malaria por plasmodium falciparum</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">vivax</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Malaria por plasmodium vivax</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t</tbody>\n</table>\n\n<p>el Segundo contempla una serie histórica del 2009 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo.</p>\n\n<p>El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $creado
#> [1] "Vie, 11/24/2023 - 08:50"
#> 
#> $modificado
#> [1] "Lun, 12/16/2024 - 12:51"
#> 
#> $recursos
#> $recursos[[1]]
#> $recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> $recursos[[1]]$formato
#> [1] "csv"
#> 
#> $recursos[[1]]$tamaño
#> [1] "5.64 MB"
#> 
#> $recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> $recursos[[1]]$modificado
#> [1] "Date changed  Lun, 12/16/2024 - 12:47"
#> 
#> 
#> $recursos[[2]]
#> $recursos[[2]]$nombre
#> [1] "datos_abiertos_vigilancia_malaria_2009_2023"
#> 
#> $recursos[[2]]$formato
#> [1] "csv"
#> 
#> $recursos[[2]]$tamaño
#> [1] "54.4 MB"
#> 
#> $recursos[[2]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2009_2023.csv"
#> 
#> $recursos[[2]]$modificado
#> [1] "Date changed  Lun, 12/16/2024 - 12:51"
#> 
#> 
#> 

# Información de todos los datasets / All datasets information
vp_dataset_info()
#> $malaria
#> $malaria$dataset
#> [1] "malaria"
#> 
#> $malaria$titulo
#> [1] "Vigilancia epidemiológica de Malaria"
#> 
#> $malaria$descripcion
#> [1] "<h1>Dataset de notificación de casos de malaria&nbsp;al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <strong>malaria</strong><b>&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483 son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>Se cuentan con dos dataset el primero contempla una serie histórica del 2000 al 2008, con datos desagregados por ano,semana,departamento,provincia,distrito,ubigeo,falciparum,vivax.</p>\n\n<p>El dataset incluye las siguientes variables:</p>\n\n<table border=\"0\" cellpadding=\"0\" cellspacing=\"0\" width=\"671\">\n\t<tbody>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\" width=\"90\">CAMPO</td>\n\t\t\t<td width=\"80\">TIPO</td>\n\t\t\t<td width=\"231\">DENOMINACIÓN</td>\n\t\t\t<td width=\"270\">VALORES</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td colspan=\"2\">Código del lugar probable de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">falciparum</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Malaria por plasmodium falciparum</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">vivax</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Malaria por plasmodium vivax</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t</tbody>\n</table>\n\n<p>el Segundo contempla una serie histórica del 2009 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo.</p>\n\n<p>El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $malaria$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $malaria$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $malaria$creado
#> [1] "Vie, 11/24/2023 - 08:50"
#> 
#> $malaria$modificado
#> [1] "Lun, 12/16/2024 - 12:51"
#> 
#> $malaria$recursos
#> $malaria$recursos[[1]]
#> $malaria$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> $malaria$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $malaria$recursos[[1]]$tamaño
#> [1] "5.64 MB"
#> 
#> $malaria$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> $malaria$recursos[[1]]$modificado
#> [1] "Date changed  Lun, 12/16/2024 - 12:47"
#> 
#> 
#> $malaria$recursos[[2]]
#> $malaria$recursos[[2]]$nombre
#> [1] "datos_abiertos_vigilancia_malaria_2009_2023"
#> 
#> $malaria$recursos[[2]]$formato
#> [1] "csv"
#> 
#> $malaria$recursos[[2]]$tamaño
#> [1] "54.4 MB"
#> 
#> $malaria$recursos[[2]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2009_2023.csv"
#> 
#> $malaria$recursos[[2]]$modificado
#> [1] "Date changed  Lun, 12/16/2024 - 12:51"
#> 
#> 
#> 
#> 
#> $dengue
#> $dengue$dataset
#> [1] "dengue"
#> 
#> $dengue$titulo
#> [1] "Vigilancia Epidemiológica de dengue"
#> 
#> $dengue$descripcion
#> [1] "<h1>Dataset de notificación de casos de dengue al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <strong>dengue</strong> al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483&nbsp;son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $dengue$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $dengue$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $dengue$creado
#> [1] "Mié, 11/22/2023 - 17:08"
#> 
#> $dengue$modificado
#> [1] "Mié, 12/11/2024 - 15:13"
#> 
#> $dengue$recursos
#> $dengue$recursos[[1]]
#> $dengue$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_dengue_2000_2023.csv"
#> 
#> $dengue$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $dengue$recursos[[1]]$tamaño
#> [1] "78.01 MB"
#> 
#> $dengue$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_dengue_2000_2023.csv"
#> 
#> $dengue$recursos[[1]]$modificado
#> [1] "Date changed  Mié, 12/11/2024 - 09:15"
#> 
#> 
#> 
#> 
#> $leishmaniasis
#> $leishmaniasis$dataset
#> [1] "leishmaniasis"
#> 
#> $leishmaniasis$titulo
#> [1] "vigilancia epidemiológica de leishmaniosis"
#> 
#> $leishmaniasis$descripcion
#> [1] "<h1>Dataset de notificación de casos de leishmaniosis&nbsp;al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <b>leishmaniosis&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483 son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $leishmaniasis$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $leishmaniasis$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $leishmaniasis$creado
#> [1] "Vie, 11/24/2023 - 08:47"
#> 
#> $leishmaniasis$modificado
#> [1] "Mié, 12/11/2024 - 14:58"
#> 
#> $leishmaniasis$recursos
#> $leishmaniasis$recursos[[1]]
#> $leishmaniasis$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_leishmaniosis_2000_2023.csv"
#> 
#> $leishmaniasis$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $leishmaniasis$recursos[[1]]$tamaño
#> [1] "15.08 MB"
#> 
#> $leishmaniasis$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_leishmaniosis_2000_2023.csv"
#> 
#> $leishmaniasis$recursos[[1]]$modificado
#> [1] "Date changed  Mié, 12/11/2024 - 14:57"
#> 
#> 
#> 
#> 
#> $zoonosis
#> $zoonosis$dataset
#> [1] "zoonosis"
#> 
#> $zoonosis$titulo
#> [1] "Vigilancia Epidemiológica de las zoonosis"
#> 
#> $zoonosis$descripcion
#> [1] "<h1>Dataset de notificación de casos de enfermedades zoonoticas al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <strong>dengue</strong> al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483&nbsp;son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $zoonosis$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $zoonosis$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $zoonosis$creado
#> [1] "Lun, 12/16/2024 - 12:57"
#> 
#> $zoonosis$modificado
#> [1] "Lun, 12/16/2024 - 12:58"
#> 
#> $zoonosis$recursos
#> $zoonosis$recursos[[1]]
#> $zoonosis$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_zoonosis_2000_2023"
#> 
#> $zoonosis$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $zoonosis$recursos[[1]]$tamaño
#> [1] "10.13 MB"
#> 
#> $zoonosis$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_zoonosis_2000_2023.csv"
#> 
#> $zoonosis$recursos[[1]]$modificado
#> [1] "Date changed  Lun, 12/16/2024 - 12:58"
#> 
#> 
#> 
#> 
#> $ira
#> $ira$dataset
#> [1] "ira"
#> 
#> $ira$titulo
#> [1] "Vigilancia epidemiologica de Infecciones Respiratoiras Agudas (IRA)"
#> 
#> $ira$descripcion
#> [1] "<h1>Dataset de notificación de casos de infecciones respiratorias agudas&nbsp;al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <b>Infecciones Respiratorias Agudas (IRA)&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232 establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483&nbsp;son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table border=\"0\" cellpadding=\"0\" cellspacing=\"0\" width=\"622\">\n\t<tbody>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\" width=\"162\">CAMPO</td>\n\t\t\t<td width=\"80\">TIPO</td>\n\t\t\t<td width=\"380\">DENOMINACIÓN</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ano</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Año de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">semana</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">sub_reg_nt</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ubigeo</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Lugar probable de infección&nbsp;(INEI)</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ira_no_neumonia</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Casos de IRAS no neumonías en menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">neumonía_&lt;5años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Episodios de neumonías en niños menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">neumonía_&gt;60 años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Episodios de neumonías en niños mayores de 60 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">Hospitalizados _&lt; 5 años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Hospitalizados por neumonías en niños menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">Hospitalizados _&gt;60 años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Hospitalizados por neumonías en niños mayores de 60 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">defunciones &lt; 5años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Defunciones por neumonías en niños menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">defunciones &gt;60 años</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Defunciones por neumonías en niños mayores de 60 años</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $ira$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $ira$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $ira$creado
#> [1] "Vie, 11/24/2023 - 08:55"
#> 
#> $ira$modificado
#> [1] "Mié, 12/11/2024 - 14:51"
#> 
#> $ira$recursos
#> $ira$recursos[[1]]
#> $ira$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_iras_2000_2023.csv"
#> 
#> $ira$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $ira$recursos[[1]]$tamaño
#> [1] "123.67 MB"
#> 
#> $ira$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_iras_2000_2023.csv"
#> 
#> $ira$recursos[[1]]$modificado
#> [1] "Date changed  Mar, 12/10/2024 - 17:30"
#> 
#> 
#> 
#> 
#> $carrion
#> $carrion$dataset
#> [1] "carrion"
#> 
#> $carrion$titulo
#> [1] "Vigilancia Epidemiológica de Enfermedad de Carrión"
#> 
#> $carrion$descripcion
#> [1] "<h1>Dataset de notificación de casos de Enfermedad de Carrion al sistema de vigilancia en salud pública del Perú</h1>\n\n<p class=\"rtejustify\">Este dataset contiene información sobre la notificación de casos de <b>Enfermedad de Carrión&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232&nbsp;establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483 son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p class=\"rtejustify\">El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table>\n\t<tbody>\n\t\t<tr>\n\t\t\t<th>CAMPO</th>\n\t\t\t<th>TIPO</th>\n\t\t\t<th>DENOMINACIÓN</th>\n\t\t\t<th>VALORES</th>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>enfermedad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Diagnóstico vigilado</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ano</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Año</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>semana</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diagnostic</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>CIE 10</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_dx</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de diagnóstico</td>\n\t\t\t<td>C= Confirmado, P = Probable, S = Sospechoso</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>diresa</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>ubigeo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Código del lugar probable de infección</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Edad del paciente</td>\n\t\t\t<td>&nbsp;</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>tipo_edad</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Tipo de edad del paciente</td>\n\t\t\t<td>A = Año, M = Mes, D = Días</td>\n\t\t</tr>\n\t\t<tr>\n\t\t\t<td>sexo</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Sexo</td>\n\t\t\t<td>M = Masculino, F = Femenino</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $carrion$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $carrion$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $carrion$creado
#> [1] "Vie, 11/24/2023 - 08:39"
#> 
#> $carrion$modificado
#> [1] "Mié, 12/11/2024 - 14:42"
#> 
#> $carrion$recursos
#> $carrion$recursos[[1]]
#> $carrion$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_enfermedad_carrion_2000_2023.csv"
#> 
#> $carrion$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $carrion$recursos[[1]]$tamaño
#> [1] "5.27 MB"
#> 
#> $carrion$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_enfermedad_carrion_2000_2023.csv"
#> 
#> $carrion$recursos[[1]]$modificado
#> [1] "Date changed  Mié, 12/11/2024 - 11:29"
#> 
#> 
#> 
#> 
#> $eda
#> $eda$dataset
#> [1] "eda"
#> 
#> $eda$titulo
#> [1] "Vigilancia epidemiológica de Enfermedad Diarreica Aguda (EDA)"
#> 
#> $eda$descripcion
#> [1] "<h1>Dataset de notificación de casos de Enfermedad&nbsp;Diarreica Aguda al sistema de vigilancia en salud pública del Perú</h1>\n\n<p>Este dataset contiene información sobre la notificación de casos de <b>Enfermedad&nbsp;Diarreica Aguda&nbsp;</b>al sistema de vigilancia en salud pública del Perú, gestionado y administrado por el <strong>Centro Nacional de Epidemiología, Prevención y Control de Enfermedades (CDC PERU)</strong>. Los datos provienen de la <strong>Red Nacional de Epidemiología (RENACE)</strong>, que está conformada por <strong>10 232 establecimientos de salud</strong> del Ministerio de Salud, EsSalud y otros del sector en los diferentes niveles de las Direcciones Regionales de Salud que tiene el Perú. De estos, <strong>8 483 son Unidades Notificantes</strong>, reconocidos con Resolución Directoral de las respectivas Diresas/Geresas de Salud del Perú. La información del dataset se actualiza cada semana o mes, según la sección de tendencia del boletín epidemiológico, y es suministrada semanalmente por la RENACE. Los datos y análisis son provisionales y pueden estar sujetos a modificación. El dataset tiene como objetivo proporcionar información sobre la situación epidemiológica del dengue en el Perú, así como facilitar el acceso a los datos abiertos para fines de investigación, educación, prevención y control de esta enfermedad.</p>\n\n<p>El dataset contempla una serie histórica del 2000 al 2023, con datos desagregados por departamento, provincia, distrito, enfermedad, año, semana, diagnóstico, tipo de diagnóstico, dirección de salud, ubigeo, edad, tipo de edad y sexo. El dataset incluye las siguientes variables:</p>\n\n<table border=\"0\" cellpadding=\"0\" cellspacing=\"0\" width=\"546\">\n\t<tbody>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\" width=\"152\">CAMPO</td>\n\t\t\t<td width=\"80\">TIPO</td>\n\t\t\t<td width=\"314\">DENOMINACIÓN</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">departamento</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Región geográfica</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">provincia</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Provincia</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">distrito</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ano</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Año de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">semana</td>\n\t\t\t<td>Caracter</td>\n\t\t\t<td>Semana de inicio de síntomas</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">sub_reg_nt</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Dirección de salud que notifica</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">ubigeo</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Lugar probable de infección</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">episodios_men5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Episodios de EDA en menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">hospitalizados_men5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Hospitalizados por EDA en menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">defunciones_men5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Defunciones por EDA en menores de 5 años</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">episodios_may5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Episodios de EDA de 5 años a más</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">hospitalizados_may5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Hospitalizados por EDA de 5 años a más</td>\n\t\t</tr>\n\t\t<tr height=\"20\">\n\t\t\t<td height=\"20\">defunciones_may5</td>\n\t\t\t<td>Integer</td>\n\t\t\t<td>Defunciones por EDA de 5 años a más</td>\n\t\t</tr>\n\t</tbody>\n</table>\n"
#> 
#> $eda$organizacion
#> [1] "Centro Nacional de Epidemiología, Prevención y Control de Enfermedades"
#> 
#> $eda$licencia
#> [1] "http://opendefinition.org/licenses/odc-by/"
#> 
#> $eda$creado
#> [1] "Vie, 11/24/2023 - 09:02"
#> 
#> $eda$modificado
#> [1] "Mié, 12/11/2024 - 14:42"
#> 
#> $eda$recursos
#> $eda$recursos[[1]]
#> $eda$recursos[[1]]$nombre
#> [1] "datos_abiertos_vigilancia_edas_200_2023.csv"
#> 
#> $eda$recursos[[1]]$formato
#> [1] "csv"
#> 
#> $eda$recursos[[1]]$tamaño
#> [1] "100.98 MB"
#> 
#> $eda$recursos[[1]]$url
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_edas_2000_2023.csv"
#> 
#> $eda$recursos[[1]]$modificado
#> [1] "Date changed  Mié, 12/11/2024 - 10:25"
#> 
#> 
#> 
#>