Obtener metadatos DCAT/DCAP-AP de un dataset / Get DCAT/DCAP-AP metadata for a dataset

Recupera los metadatos completos de un dataset en formato DCAT/DCAP-AP, incluyendo información sobre recursos, distribuciones y términos de uso.

Retrieves complete dataset metadata in DCAT/DCAP-AP format, including information about resources, distributions and terms of use.

Uso

vp_metadata(dataset, refresh = FALSE)

Argumentos

dataset

Nombre del dataset ("malaria", "dengue", etc.) / Dataset name ("malaria", "dengue", etc.)

refresh

Lógico. Si TRUE, actualiza la información desde el portal. / Logical. If TRUE, refreshes information from the portal.

Valor

Lista con metadatos DCAT / List with DCAT metadata

Ejemplos

# Obtener metadatos de malaria / Get malaria metadata
metadata <- vp_metadata("malaria")

# Ver términos DCAT disponibles / View available DCAT terms
names(metadata)
#>  [1] "@context"          "@type"             "dct:identifier"   
#>  [4] "dct:title"         "dct:description"   "dct:issued"       
#>  [7] "dct:modified"      "dcat:landingPage"  "dct:license"      
#> [10] "dct:publisher"     "dcat:theme"        "dcat:distribution"
#> [13] "dcat:keyword"      "dct:spatial"       "dct:temporal"     
#> [16] "_raw"             

# Acceder a distribuciones / Access distributions
metadata$`dcat:distribution`
#> [[1]]
#> [[1]]$`@type`
#> [1] "dcat:Distribution"
#> 
#> [[1]]$`dct:identifier`
#> [1] "a1ddd151-59aa-49f8-9c20-5e9a3eca9307"
#> 
#> [[1]]$`dct:title`
#> [1] "datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> [[1]]$`dct:description`
#> [1] ""
#> 
#> [[1]]$`dcat:accessURL`
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> [[1]]$`dcat:downloadURL`
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2000_2008.csv"
#> 
#> [[1]]$`dct:format`
#> [1] "csv"
#> 
#> [[1]]$`dcat:byteSize`
#> [1] 5913968
#> 
#> [[1]]$`dct:issued`
#> [1] "Vie, 11/24/2023 - 08:52"
#> 
#> [[1]]$`dct:modified`
#> [1] "Date changed  Lun, 12/16/2024 - 12:47"
#> 
#> 
#> [[2]]
#> [[2]]$`@type`
#> [1] "dcat:Distribution"
#> 
#> [[2]]$`dct:identifier`
#> [1] "43e48a55-0744-4d1b-b6f5-c8a2c83af386"
#> 
#> [[2]]$`dct:title`
#> [1] "datos_abiertos_vigilancia_malaria_2009_2023"
#> 
#> [[2]]$`dct:description`
#> [1] ""
#> 
#> [[2]]$`dcat:accessURL`
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2009_2023.csv"
#> 
#> [[2]]$`dcat:downloadURL`
#> [1] "https://www.datosabiertos.gob.pe/sites/default/files/datos_abiertos_vigilancia_malaria_2009_2023.csv"
#> 
#> [[2]]$`dct:format`
#> [1] "csv"
#> 
#> [[2]]$`dcat:byteSize`
#> [1] 57042534
#> 
#> [[2]]$`dct:issued`
#> [1] "Lun, 12/16/2024 - 12:51"
#> 
#> [[2]]$`dct:modified`
#> [1] "Date changed  Lun, 12/16/2024 - 12:51"
#> 
#>